+++ test/rules/prefer_entry_point_imports/examples/example.dart
@@ -1,0 +1,4 @@
+import 'dart:io';
+import 'package:dart_code_metrics/src/rules/prefer_entry_point_imports.dart';
+import '../../../../lib/src/rules/prefer_entry_point_imports.dart';
+import 'package:test/test.dart';